Refactored code and splitted it across more files with a directory structure
