+++ popup/index.html
@@ -64,7 +64,7 @@
   <!-- JS sources -->
   <script src="./lib/fuse.min.js"></script>
   <script src="./lib/flexsearch.min.js"></script>
-  <script src="./js/search.js" type="module"></script>
+  <script src="./js/initSearch.js" type="module"></script>
   <script src="./lib/tagify.min.js"></script>
   <script src="./lib/mark.min.js"></script>
 </body>
