+++ popup/js/helper/browserApi.js
