+++ popup/js/helper/utils.js
