+++ popup/js/search/flexSearch.js
