+++ popup/options.html
@@ -45,7 +45,7 @@
 
   <!-- JS sources -->
   <script src="./lib/json5.min.js"></script>
-  <script src="./js/editOptions.js" type="module"></script>
+  <script src="./js/initOptions.js" type="module"></script>
 </body>
 
 </html>
